refactor(core): pi_hooks_move folds to four files (KEN-548) [no-changelog]
Ten files with super:: ping-pong become four with no imports between
siblings except preflight -> migrated and disposal -> migrated. Bodies
move byte-for-byte; the diff is mod, use, visibility and doc-comment
lines. No behavior change.

- {preflight,holding,claims,registry}.rs -> preflight.rs: one file, since
  claims and holding each need Preflight and holding needs claims, so no
  seam among them stands alone
- retire.rs -> mod.rs
- {identity,record}.rs -> migrated.rs: every reader of Identity and
  registered() is a registry reading, and the record is the other half of
  "has the move finished"
- disposal.rs stays; it now names super::migrated::Identity directly

The three merged rows rise under RATCHET_RAISE=1 (size-ratchet: merging
fragments). No row existed for any of the ten before, so none is deleted.
Items that were pub(super) for a sibling now in the same file are private;
Held stays pub(super) because it is in claim()'s signature, which mod.rs
calls.
